finished a preliminary implementation of back disassemble.  there are some kinks
moar banks for gb

Fix some startup valgrind warnings

Better x86 analysis and 'ao' now accepts a counter argument

Fix malloc://-1

Minor stuff

Fix of scr.color in disasm.c

Fix #389 - Fix segfault in ptrace://-1

Fix ranal2 build, remove ningb debug-printf

Fix libr plugins path

Fix #388 - yank length field segfault

Fix 'cg /' crash and only use numeric values in web yank (thanks zlul!)

Add some misc opcodes and remove duplicates

Fixes in tcp connect return value and implement 2> pipes

Add SNES disassembler

Fix build

Fix ^w segfault and many bugfixes for macro scripting

Fix #382 - Refactor 16bit addressing issue to fix rel32 calculation

Fixed some merge blunders

+++ libr/io/p/io_malloc.c
@@ -96,7 +96,7 @@ static RIODesc *__open(RIO *io, const char *pathname, int rw, int mode) {
 		} else {
 			mal->size = r_num_math (NULL, pathname+9);
 			mal->offset = 0;
-			if ((mal->size)>0) {
+			if (((int)(mal->size))>0) {
 				mal->buf = malloc (mal->size);
 				memset (mal->buf, '\0', mal->size);
 			} else {
